# api/openapi.yaml
openapi: 3.1.0
...
info:
  title: JobTrack API
  version: 1.0.0
  description: |
    Job search for software engineers. Postings are read from employers' own
    applicant tracking systems, scored against a user profile, and tracked
    through an application funnel.

    Two conventions run through every response:

    - `null` means NOT KNOWN and is never collapsed to a zero or an empty
      string. An undisclosed salary is `null`, not `0`; the difference between
      "the employer did not publish this" and "this role pays nothing" is the
      whole point.
    - Errors are RFC 9457 problem+json, always, including 404s from unmatched
      routes.
servers:
  - url: /
    description: Same-origin. The browser never talks to the API cross-origin,
      so the session cookie needs neither CORS nor SameSite=None.
...
paths:
...
  /v1/me/preferences:
    get:
      tags: [profile]
      summary: Display preferences
      operationId: getPreferences
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Preferences' }
        '401': { $ref: '#/components/responses/Problem' }
    patch:
      tags: [profile]
      summary: Update display preferences
      operationId: updatePreferences
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/Preferences' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Preferences' }
        '400': { $ref: '#/components/responses/Problem' }
...
components:
  responses:
    Problem:
      description: RFC 9457 problem detail
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/ProblemDetail' }
  schemas:
...
    Preferences:
      type: object
      description: |
        Display settings, stored as JSONB on the user.

        There is no `density`. One existed and was removed: it changed row
        padding without changing type size or target size, so it did not do the
        accessibility job it claimed to. Rows written before the removal still
        carry the key and it is preserved untouched, but nothing reads it.
      properties:
        theme: { type: string, enum: [system, light, dark] }
...
    FieldError:
      type: object
      required: [field, code, message]
      properties:
        field: { type: string }
        code: { type: string }
        message: { type: string }
    ProblemDetail:
      type: object
      description: RFC 9457.
      required: [type, title, status]
      properties:
        type: { type: string, format: uri }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        instance: { type: string }
        fields: { type: array, items: { $ref: '#/components/schemas/FieldError' } }
        trace_id:
          type: string
          description: Correlates a user-visible error with its server logs and traces.